# backend/openapi.yml
openapi: 3.0.0
info:
  title: Analysis Results API
  version: 1.0.0
  description: API for submitting sequences and retrieving their processing results.
servers:
  - url: http://localhost:8080
    description: Local server for backend

paths:
  /api/analyze:
    post:
      summary: Submit sequences for analysis
      requestBody:
        description: Submit sequences for comparison and analysis
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sequence_wt:
                  type: string
                  description: Wild-type sequence
                  example: "AUGGCCUAA"
                sequence_mt:
                  type: string
                  description: Mutant sequence
                  example: "AUGGCCGUA"
              required:
                - sequence_wt
                - sequence_mt
      responses:
        '200':
          description: Analysis successfully submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Unique task identifier
                    example: "123e4567-e89b-12d3-a456-426614174000"
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid input format"

  /api/status/{id}:
    get:
      summary: Check the status of a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: Unique task identifier
      responses:
        '200':
          description: Task status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Current task status
                    example: "IN_PROGRESS"
                  reason:
                    type: string
                    nullable: true
                    description: Reason for failure, if applicable
                    example: null
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid task id"

  /api/result/{id}:
    get:
      summary: Retrieve the result of a completed task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: Unique task identifier
      responses:
        '200':
          description: Task results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  RNApdist:
                    type: number
                    format: float
                    description: RNApdist score
                    example: 0.334
                  RNAdistance:
                    type: object
                    properties:
                      f:
                        type: integer
                        description: First distance measure
                        example: 16
                      c:
                        type: integer
                        description: Second distance measure
                        example: 2
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid task id"

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        status:
          type: string
          example: "IN_PROGRESS"
        result:
          type: object
          properties:
            RNApdist:
              type: number
              format: float
              example: 0.334
            RNAdistance:
              type: object
              properties:
                f:
                  type: integer
                  example: 16
                c:
                  type: integer
                  example: 2

# database swagger spec

paths:
  /test-db:
    get:
      summary: Test database connection
      responses:
        '200':
          description: Successfully connected to the database.
          content:
            application/json:
              schema:
                type: object
                properties:
                  DATABASE():
                    type: string
                    description: Name of the connected database.
        '500':
          description: Failed to connect to the database.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message.

  /analysis-results:
    post:
      summary: Add a new analysis result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                normalSequence:
                  type: string
                  description: Normal sequence being analyzed.
                wildSequence:
                  type: string
                  description: Wild sequence being analyzed.
                result:
                  type: string
                  description: Result of the analysis.
              required:
                - normalSequence
                - wildSequence
                - result
      responses:
        '201':
          description: Analysis result added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Result added successfully!"
        '500':
          description: Failed to connect to the database.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to connect to the database"

    get:
      summary: Retrieve all analysis results
      responses:
        '200':
          description: A list of all analysis results.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: ID of the result.
                    normalSequence:
                      type: string
                      description: Normal sequence being analyzed.
                    wildSequence:
                      type: string
                      description: Wild sequence being analyzed.
                    result:
                      type: string
                      description: Result of the analysis.
        '500':
          description: Failed to connect to the database.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message.

  /add-analysis-result:
    get:
      summary: Add a new analysis result using query parameters
      parameters:
        - in: query
          name: normalSequence
          schema:
            type: string
          required: true
          description: Normal sequence being analyzed.
        - in: query
          name: wildSequence
          schema:
            type: string
          required: true
          description: Wild sequence being analyzed.
        - in: query
          name: result
          schema:
            type: string
          required: true
          description: Result of the analysis.
      responses:
        '201':
          description: Analysis result added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Result added successfully!"
        '400':
          description: Missing parameters in the request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing parameters"
        '500':
          description: Failed to connect to the database.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to connect to the database"

components:
  schemas:
    AnalysisResult:
      type: object
      properties:
        normalSequence:
          type: string
          description: Normal sequence being analyzed.
        wildSequence:
          type: string
          description: Wild sequence being analyzed.
        result:
          type: string
          description: Result of the analysis.
      required:
        - normalSequence
        - wildSequence
        - result